Refactor type usage in AGENTS.md and related files for 64-bit safety
- Updated AGENTS.md to document type usage guidelines, emphasizing the distinction between `Int`, `ssize_t`, and `CLogicInt` for new code.
- Modified `CGrepAgent.cpp` to replace `int` with `size_t` for path iteration, enhancing compatibility with larger datasets.
- Adjusted `CNativeW.cpp` to compare string lengths using `ssize_t` and `size_t`, ensuring accurate comparisons without data loss.
- Improved overall consistency in type handling across the codebase to support 64-bit architecture.

+++ sakura_core/agent/CGrepAgent.cpp
@@ -768,7 +768,7 @@ DWORD CGrepAgent::DoGrep(
 		}
 		nHitCount = nGrepTreeResult;
 	}else{
-		for( int nPath = 0; nPath < (int)vPaths.size(); nPath++ ){
+		for( size_t nPath = 0; nPath < vPaths.size(); nPath++ ){
 			bool bOutputBaseFolder = false;
 			std::wstring sPath = ChopYen( vPaths[nPath] );
 			int nTreeRet = DoGrepTree(
